+++ clang-tools-extra/test/clang-tidy/checkers/readability/inconsistent-ifelse-braces.cpp
@@ -1,4 +1,4 @@
-// RUN: %check_clang_tidy -std=c++98-or-later %s bugprone-inconsistent-ifelse-braces %t
+// RUN: %check_clang_tidy -std=c++98-or-later %s readability-inconsistent-ifelse-braces %t
 
 bool cond(const char *) { return false; }
 void do_something(const char *) {}
@@ -8,13 +8,13 @@ void f() {
   if (cond("if0") /*comment*/) do_something("if-same-line");
   else {
   }
-  // CHECK-MESSAGES: :[[@LINE-3]]:31: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-3]]:31: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: if (cond("if0") /*comment*/) { do_something("if-same-line");
   // CHECK-FIXES: } else {
 
   if (cond("if0.1") /*comment*/) {
   } else do_something("else-same-line");
-  // CHECK-MESSAGES: :[[@LINE-1]]:9: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-1]]:9: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: } else { do_something("else-same-line");
   // CHECK-FIXES: }
 
@@ -22,14 +22,14 @@ void f() {
     do_something("if-single-line");
   else {
   }
-  // CHECK-MESSAGES: :[[@LINE-4]]:19: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-4]]:19: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: if (cond("if1")) {
   // CHECK-FIXES: } else {
 
   if (cond("if1.1")) {
   } else
     do_something("else-single-line");
-  // CHECK-MESSAGES: :[[@LINE-2]]:9: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-2]]:9: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: } else {
   // CHECK-FIXES: }
 
@@ -38,7 +38,7 @@ void f() {
     do_something("if-multi-line");
   else {
   }
-  // CHECK-MESSAGES: :[[@LINE-5]]:31: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-5]]:31: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: if (cond("if2") /*comment*/) {
   // CHECK-FIXES: } else {
 
@@ -46,7 +46,7 @@ void f() {
   } else
     // some comment
     do_something("else-multi-line");
-  // CHECK-MESSAGES: :[[@LINE-3]]:9: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-3]]:9: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: } else {
   // CHECK-FIXES: }
 
@@ -54,7 +54,7 @@ void f() {
   else if (cond("if3")) {
   } else {
   }
-  // CHECK-MESSAGES: :[[@LINE-4]]:19: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-4]]:19: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: if (cond("if3")) { do_something("elseif-same-line");
   // CHECK-FIXES: } else if (cond("if3")) {
 
@@ -62,14 +62,14 @@ void f() {
   } else if (cond("if3.1")) do_something("elseif-same-line");
   else {
   }
-  // CHECK-MESSAGES: :[[@LINE-3]]:28: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-3]]:28: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: } else if (cond("if3.1")) { do_something("elseif-same-line");
   // CHECK-FIXES: } else {
 
   if (cond("if3.2")) {
   } else if (cond("if3.2")) {
   } else do_something("else-same-line");
-  // CHECK-MESSAGES: :[[@LINE-1]]:9: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-1]]:9: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: } else { do_something("else-same-line");
   // CHECK-FIXES: }
 
@@ -80,8 +80,8 @@ void f() {
     }
   else {
   }
-  // CHECK-MESSAGES: :[[@LINE-7]]:25: warning: <message> [bugprone-inconsistent-ifelse-braces]
-  // CHECK-MESSAGES: :[[@LINE-7]]:27: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-7]]:25: warning: <message> [readability-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-7]]:27: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: if (cond("if4-outer")) {
   // CHECK-FIXES: if (cond("if4-inner")) {
   // CHECK-FIXES: } else {
@@ -91,14 +91,14 @@ void f() {
       do_something("if-single-line");
   else if (cond("if5")) {
   }
-  // CHECK-MESSAGES: :[[@LINE-4]]:19: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-4]]:19: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: if (cond("if5")) {
   // CHECK-FIXES: } else if (cond("if5")) {
 
   if (cond("if5.1")) {
   } else if (cond("if5.1"))
       do_something("elseif-single-line");
-  // CHECK-MESSAGES: :[[@LINE-2]]:28: warning: <message> [bugprone-inconsistent-ifelse-braces]
+  // CHECK-MESSAGES: :[[@LINE-2]]:28: warning: <message> [readability-inconsistent-ifelse-braces]
   // CHECK-FIXES: } else if (cond("if5.1")) {
   // CHECK-FIXES: }
 }