+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -184,8 +184,8 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
-- New :doc:`bugprone-inconsistent-ifelse-braces
-  <clang-tidy/checks/bugprone/inconsistent-ifelse-braces>` check.
+- New :doc:`readability-inconsistent-ifelse-braces
+  <clang-tidy/checks/readability/inconsistent-ifelse-braces>` check.
 
   Detects ``if``/``else`` statements where one branch uses braces and the other
   does not.